chore(External Config): OZF-863 Create external OzoneConfig.js file
Configurations are now stored in a seperate OzoneConfig.js. The source
of this file is in the `app` directory.  The built version resides in
the top level of the distribution.  At build-time the config file is
constructed using environment variables along with default values.

Breaking changes:
LOGOUT_URL environment variable has been removed as it was no longer used
HUD_URL environment variable is now set as part of the buildGhPages task